+++ deploy/complete/terraform/outputs.tf
@@ -21,7 +21,7 @@ output "mushop_url_https" {
   depends_on = [helm_release.ingress_nginx]
 }
 output "grafana_url" {
-  value       = var.grafana_enabled ? format("${local.mushop_url_protocol}://%s/grafana", local.mushop_ingress_hostname) : null 
+  value       = var.grafana_enabled ? format("${local.mushop_url_protocol}://%s/grafana", local.mushop_ingress_hostname) : null
   description = "Grafana Dashboards URL"
 
   depends_on = [helm_release.ingress_nginx]
@@ -41,18 +41,15 @@ output "autonomous_database_password" {
   value = random_string.autonomous_database_admin_password.result
 }
 output "grafana_admin_password" {
-  value = var.grafana_enabled ? nonsensitive(data.kubernetes_secret.mushop_utils_grafana.0.data.admin-password) : null # Required for TF 0.15, as automatically generate an error if is not marked as sensitive
-  # value     = data.kubernetes_secret.mushop_utils_grafana.data.admin-password # TF 0.14 version support by ORM does not support nonsensitive function
+  value = var.grafana_enabled ? nonsensitive(local.grafana_admin_password) : null # Required for TF 0.15, as automatically generate an error if is not marked as sensitive
+  # value     = local.grafana_admin_password # TF 0.14 version support by ORM does not support nonsensitive function
 }
 output "mushop_source_code" {
   value = "https://github.com/oracle-quickstart/oci-cloudnative/"
 }
 locals {
-  mushop_ingress_ip       = data.kubernetes_service.mushop_ingress.load_balancer_ingress[0].ip
-  mushop_ingress_hostname = var.ingress_hosts == "" ? data.kubernetes_service.mushop_ingress.load_balancer_ingress[0].ip : split(",", var.ingress_hosts)[0]
+  mushop_ingress_ip       = var.ingress_nginx_enabled ? data.kubernetes_service.mushop_ingress.0.load_balancer_ingress.0.ip : "#Ingress_Not_Deployed"
+  mushop_ingress_hostname = var.ingress_nginx_enabled ? (var.ingress_hosts == "" ? data.kubernetes_service.mushop_ingress.0.load_balancer_ingress.0.ip : split(",", var.ingress_hosts)[0]) : "#Ingress_Not_Deployed"
   mushop_url_protocol     = var.ingress_tls ? "https" : "http"
-}
-
-output "temp" {
-  value = oci_apigateway_deployment.newsletter_subscription.endpoint
+  grafana_admin_password  = var.grafana_enabled ? data.kubernetes_secret.mushop_utils_grafana.0.data.admin-password : "Grafana_Not_Deployed"
 }