+++ src/storefront/gulpfile.js
@@ -107,9 +107,9 @@ gulp.task('image:copy', function() {
 gulp.task('image:optimize', function(done) {
   return gulp.src(`${opt.buildDir}/images/**/*`)
     .pipe(cache(imagemin([
-      pngquant(),
+      pngquant({quality: [0.3, 0.5]}),
       imagemin.gifsicle({interlaced: true}),
-      imagemin.mozjpeg({progressive: true, quality: 85}),
+      imagemin.mozjpeg({progressive: true, quality: 55}),
       imagemin.svgo({
         plugins: [
           { removeViewBox: true },